// frontend/src/api-client/model/spot-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SpotReservation.Api | v1
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface SpotDto {
    'id': string;
    'name': string;
    'description': string | null;
    'isActive': boolean;
    'pricePerDay': number;
    'createdAtUtc': string;
    'location': any | null;
    'pageId': string;
    'occupiedDates': Array<string>;
}